исправлены юнит тесты
