+++ src/main/java/ru/kampaii/examples/repositories/servises/UserService.java
@@ -2,6 +2,8 @@ package ru.kampaii.examples.repositories.servises;
 
 import ru.kampaii.examples.domain.entities.UsersEntity;
 
+import java.sql.SQLException;
+
 public interface UserService {
-    UsersEntity createUser(String name, Integer numOfAcc);
+    UsersEntity createUser(String name, Integer numOfAcc) throws SQLException;
 }
