+++ src/test/ru/kampaii/examples/repositories/UserServicesTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import ru.kampaii.examples.config.DatabaseConnectorProvider;
 import ru.kampaii.examples.domain.entities.AccountsEntity;
@@ -13,12 +12,12 @@ import ru.kampaii.examples.repositories.id.generators.PooledIdGeneratorImpl;
 import ru.kampaii.examples.repositories.servises.UserService;
 import ru.kampaii.examples.repositories.servises.UserServiceCommon;
 import ru.kampaii.examples.repositories.servises.UserServiceTransactional;
+import ru.kampaii.examples.repositories.servises.UserServiceTransactionalBatch;
 
 import java.sql.Connection;
 import java.sql.SQLException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
 
 @ExtendWith(MockitoExtension.class)
 public class UserServicesTest {
@@ -54,22 +53,14 @@ public class UserServicesTest {
     }
 
     @Test
-    void userServiceUnitTest() throws SQLException {
-        usersRepository = Mockito.mock(UsersRepositoryImpl.class);
-        accountsRepository = Mockito.mock(AccountsRepositoryImpl.class);
-        UserService service = new UserServiceCommon(usersRepository, accountsRepository);
-        executeUnitTest(service);
+    void userServiceTransBatchTest() throws SQLException {
+        usersRepository = new UsersRepositoryPreparedImpl(connection, new PooledIdGeneratorImpl(connection, "users", "id", 1, 1000));
+        accountsRepository = new AccountsRepositoryPreparedImpl(connection, new PooledIdGeneratorImpl(connection, "accounts", "number", 1, 1000));
+        UserService service = new UserServiceTransactionalBatch(usersRepository, accountsRepository, connection);
+        executeIntegrationTest(service);
     }
 
-    @Test
-    void userServiceTransUnitTest() throws SQLException {
-        usersRepository = Mockito.mock(UsersRepositoryImpl.class);
-        accountsRepository = Mockito.mock(AccountsRepositoryImpl.class);
-        UserService service = new UserServiceTransactional(usersRepository, accountsRepository, connection);
-        executeUnitTest(service);
-    }
-
-    void executeIntegrationTest(UserService service) {
+    void executeIntegrationTest(UserService service) throws SQLException {
         String name = "TEST";
         int firstAccountsCount = accountsRepository.count();
         int firstUsersCount = usersRepository.count();
@@ -79,11 +70,4 @@ public class UserServicesTest {
         assertEquals(1, usersRepository.count() - firstUsersCount);
     }
 
-    void executeUnitTest(UserService service) throws SQLException {
-        String name = "TEST";
-        Mockito.when(usersRepository.create(any())).thenReturn(new UsersEntity(1, name, 0F));
-        Mockito.when(accountsRepository.create(any())).thenReturn(new AccountsEntity(1, 0F, 1, 1));
-        UsersEntity entity = service.createUser(name, 5);
-        assertEquals(name, entity.getName());
-    }
 }
